refactor: logging

+++ cmd/ableton-patcher/interactions.go
@@ -16,6 +16,7 @@ func (app *application) mainMenu() {
 		fmt.Print("1. Patch\n2. Unpatch\n3. Deauthorize\n4. Generate license\n5. Generate DSA key pair")
 		fmt.Print("\nSelect option: ")
 		option := GetLine()
+		ClearScreen()
 		switch option {
 		case "1":
 			app.patcher(false)
@@ -28,22 +29,23 @@ func (app *application) mainMenu() {
 		case "5":
 			app.keyGenerator()
 		default:
-			fmt.Print("Invalid option\n\n")
-			continue
+			fmt.Print("invalid option")
 		}
+		Pause()
 	}
 }
 
 func (app *application) deauthorize() {
 	selectedInstsData := installationDataSelector()
 
-	for _, data := range selectedInstsData {
+	for i, data := range selectedInstsData {
 		unlockFile := filepath.Join(data.Path, "/Unlock/Unlock.json")
 		os.Remove(unlockFile)
-		fmt.Printf("\nDeauthorized (%s)", data.Name)
+		fmt.Printf("Deauthorized (%s)", data.Name)
+		if i != len(selectedInstsData)-1 {
+			fmt.Print("\n")
+		}
 	}
-
-	fmt.Print("\n\n")
 }
 
 func (app *application) licenseGenerator() {
@@ -58,6 +60,7 @@ func (app *application) licenseGenerator() {
 	}
 
 	var edition int
+	var editionName string
 	fmt.Print("1. Suite\n2. Standard\n3. Intro\n4. Lite")
 	for {
 		fmt.Print("\nSelect edition: ")
@@ -65,12 +68,16 @@ func (app *application) licenseGenerator() {
 		switch editionNumber {
 		case "1":
 			edition = 2
+			editionName = "Suite"
 		case "2":
 			edition = 0
+			editionName = "Standard"
 		case "3":
 			edition = 3
+			editionName = "Intro"
 		case "4":
 			edition = 4
+			editionName = "Lite"
 		default:
 			fmt.Println("Invalid edition")
 			continue
@@ -83,11 +90,17 @@ func (app *application) licenseGenerator() {
 		LogFatalError("generate license", err)
 	}
 
-	if err := ableton.WriteAuthorizationFile(license, fmt.Sprintf("Authorize_%s.auz", hwid)); err != nil {
+	filename := fmt.Sprintf("Authorize_%s_%s.auz", editionName, hwid)
+	path, err := ExecutableDirFilePath(filename)
+	if err != nil {
+		LogFatalError("get executable path", err)
+	}
+
+	if err := ableton.WriteAuthorizationFile(license, path); err != nil {
 		LogFatalError("write authorization file", err)
 	}
 
-	fmt.Print("Done\n\n")
+	fmt.Print("License file generated")
 }
 
 func (app *application) keyGenerator() {
@@ -125,7 +138,7 @@ func (app *application) keyGenerator() {
 		if err != nil {
 			LogFatalError("save config", err)
 		}
-		fmt.Print("Config saved\n\n")
+		fmt.Print("Config saved")
 	}
 }
 
@@ -141,28 +154,34 @@ func (app *application) patcher(reverse bool) {
 
 	selectedInsts := installationSelector()
 
-	for _, installation := range selectedInsts {
+	for i, installation := range selectedInsts {
 		fileBytes, err := os.ReadFile(installation.Path)
 		if err != nil {
 			LogFatalError("open file", err)
 		}
 
 		var newContents string
+		var replacementsCounter int
 		if reverse {
-			newContents = strings.Replace(string(fileBytes), public, app.config.OriginalPublicKey, -1)
+			newContents, replacementsCounter = ReplaceAndCount(string(fileBytes), public, app.config.OriginalPublicKey)
 		} else {
-			newContents = strings.Replace(string(fileBytes), app.config.OriginalPublicKey, public, -1)
+			newContents, replacementsCounter = ReplaceAndCount(string(fileBytes), app.config.OriginalPublicKey, public)
+		}
+		if replacementsCounter == 0 {
+			fmt.Printf("Key not found (%s)", installation.Path)
+		} else {
+			err = os.WriteFile(installation.Path, []byte(newContents), 0)
+			if err != nil {
+				LogFatalError("write file", err)
+			}
+
+			fmt.Printf("Patch/Unpatch applied succesfully (%s)", installation.Path)
 		}
 
-		err = os.WriteFile(installation.Path, []byte(newContents), 0)
-		if err != nil {
-			LogFatalError("write file", err)
+		if i != len(selectedInsts)-1 {
+			fmt.Print("\n")
 		}
-
-		fmt.Printf("\nPatch/Unpatch applied (%s)", installation.Path)
 	}
-
-	fmt.Print("\n\n")
 }
 
 func installationSelector() []ableton.Installation {
@@ -171,20 +190,15 @@ func installationSelector() []ableton.Installation {
 		LogFatalError("find installations", err)
 	}
 
-	fmt.Println()
+	installations = append(installations, ableton.Installation{Path: "", Name: "Custom"})
 	installationsLen := len(installations)
-	if installationsLen == 0 {
-		fmt.Print("\nNo installations were found, please enter the path to the binary file: ")
-		input := GetLine()
-		_, err = os.Stat(input)
-		if err != nil {
-			LogFatalError("installation path not found", err)
-		}
-		installations = append(installations, ableton.Installation{Path: input, Name: "Custom Installation"})
-	}
 
 	for i, installation := range installations {
-		fmt.Printf("%d. %s (%s)\n", i+1, installation.Name, installation.Path)
+		if installation.Path != "" {
+			fmt.Printf("%d. %s (%s)\n", i+1, installation.Name, installation.Path)
+		} else {
+			fmt.Printf("%d. %s\n", i+1, installation.Name)
+		}
 	}
 
 	var selectedInsts []ableton.Installation
@@ -195,13 +209,22 @@ func installationSelector() []ableton.Installation {
 		for _, result := range requestedInsts {
 			resultInt, err := strconv.Atoi(result)
 			if err != nil {
-				fmt.Printf("invalid installation number: %s\n", result)
+				fmt.Printf("invalid installation number: %s", result)
 				continue
 			}
 
 			if resultInt > installationsLen || resultInt == 0 {
-				fmt.Printf("invalid installation number: %d\n\n", resultInt)
+				fmt.Printf("invalid installation number: %d", resultInt)
 				continue
+			}
+
+			if installations[resultInt-1].Path == "" {
+				fmt.Print("Enter path to the Live executable: ")
+				installations[resultInt-1].Path = GetLine()
+				if _, err := os.Stat(installations[resultInt-1].Path); err != nil {
+					fmt.Printf("invalid executable path")
+					continue
+				}
 			}
 
 			selectedInsts = append(selectedInsts, installations[resultInt-1])
@@ -218,20 +241,14 @@ func installationDataSelector() []ableton.InstallationData {
 		LogFatalError("find installation data", err)
 	}
 
-	fmt.Println()
 	dataLen := len(data)
-	if dataLen == 0 {
-		fmt.Print("No installation data was found, please enter the path to the data directory: ")
-		input := GetLine()
-		_, err = os.Stat(input)
-		if err != nil {
-			LogFatalError("installation data directory does not exists", err)
-		}
-		data = append(data, ableton.InstallationData{Path: input, Name: "Custom Installation Data"})
-	}
 
 	for i, installationData := range data {
-		fmt.Printf("%d. %s (%s)\n", i+1, installationData.Name, installationData.Path)
+		if installationData.Path != "" {
+			fmt.Printf("%d. %s (%s)\n", i+1, installationData.Name, installationData.Path)
+		} else {
+			fmt.Printf("%d. %s\n", i+1, installationData.Name)
+		}
 	}
 
 	var selectedInstsData []ableton.InstallationData
@@ -249,6 +266,15 @@ func installationDataSelector() []ableton.InstallationData {
 			if resultInt > dataLen || resultInt == 0 {
 				fmt.Printf("invalid installation data number: %d", resultInt)
 				continue
+			}
+
+			if data[resultInt-1].Path == "" {
+				fmt.Print("Enter path to the Live data folder: ")
+				data[resultInt-1].Path = GetLine()
+				if _, err := os.Stat(data[resultInt-1].Path); err != nil {
+					fmt.Printf("invalid data folder path")
+					continue
+				}
 			}
 
 			selectedInstsData = append(selectedInstsData, data[resultInt-1])
